Unpredictable String without testing See detail
Because testing up to this point has been manual only.  We learn how to
do doctest our code.  This gives us the advantage of being to reproduce
easily our test runs.

This is not something that is generally used in the Ruby Community.
However, it models the testing done in IRB, which I believe it is very
likely everyone one does.

UnpredictableString initially and by specification had the problem of
inheriting from String.  We learn that String class does not call to_s
when it is representing its own objects.  So Kernel#puts which generally
calls an objects to_s method does not do so when the object inherits
from String.

Rubinius does not have the same 'feature'.  So this could be designed
the same was but could still inherit from String.